+++ tests/compile_fails/invariant_dep_extraction.rs
@@ -17,7 +17,7 @@ impl Owner for InvarOwner {
     fn make_dependent(
         &self,
         (): Self::Context<'_>,
-    ) -> Result<<Self as HasDependent<'_>>::Dependent, Self::Error> {
+    ) -> Result<Dependent<'_, Self>, Self::Error> {
         Ok(PhantomData)
     }
 }