+++ tests/compile_fails/not_send_owner.rs
@@ -2,7 +2,7 @@ extern crate pair;
 
 use std::{convert::Infallible, sync::MutexGuard};
 
-use pair::{HasDependent, Owner, Pair};
+use pair::{Dependent, HasDependent, Owner, Pair};
 
 struct NotSend(MutexGuard<'static, ()>);
 
@@ -17,7 +17,7 @@ impl Owner for NotSend {
     fn make_dependent<'owner>(
         &'owner self,
         (): Self::Context<'_>,
-    ) -> Result<<Self as HasDependent<'owner>>::Dependent, Self::Error> {
+    ) -> Result<Dependent<'owner, Self>, Self::Error> {
         unimplemented!()
     }
 }
@@ -26,8 +26,8 @@ fn check_send<T: Send>() {}
 fn check_sync<T: Sync>() {}
 
 fn main() {
-    check_send::<<NotSend as HasDependent>::Dependent>();
-    check_sync::<(NotSend, <NotSend as HasDependent>::Dependent)>();
+    check_send::<Dependent<'_, NotSend>>();
+    check_sync::<(NotSend, Dependent<'_, NotSend>)>();
 
     check_send::<Pair<NotSend>>();
     check_sync::<Pair<NotSend>>();