+++ tests/compile_fails/not_sync_owner.rs
@@ -2,7 +2,7 @@ extern crate pair;
 
 use std::{cell::UnsafeCell, convert::Infallible};
 
-use pair::{HasDependent, Owner, Pair};
+use pair::{Dependent, HasDependent, Owner, Pair};
 
 struct NotSync(UnsafeCell<()>);
 
@@ -17,7 +17,7 @@ impl Owner for NotSync {
     fn make_dependent<'owner>(
         &'owner self,
         (): Self::Context<'_>,
-    ) -> Result<<Self as HasDependent<'owner>>::Dependent, Self::Error> {
+    ) -> Result<Dependent<'owner, Self>, Self::Error> {
         unimplemented!()
     }
 }
@@ -26,8 +26,8 @@ fn check_send<T: Send>() {}
 fn check_sync<T: Sync>() {}
 
 fn main() {
-    check_send::<(NotSync, <NotSync as HasDependent>::Dependent)>();
-    check_sync::<<NotSync as HasDependent>::Dependent>();
+    check_send::<(NotSync, Dependent<'_, NotSync>)>();
+    check_sync::<Dependent<'_, NotSync>>();
 
     check_send::<Pair<NotSync>>();
     check_sync::<Pair<NotSync>>();