+++ tests/compile_fails/pair_not_covariant.rs
@@ -22,7 +22,7 @@ impl Owner for Foo<'_> {
     fn make_dependent(
         &self,
         (): Self::Context<'_>,
-    ) -> Result<<Self as HasDependent<'_>>::Dependent, Self::Error> {
+    ) -> Result<Dependent<'_, Self>, Self::Error> {
         unimplemented!()
     }
 }