+++ tests/nested_pair.rs
@@ -1,6 +1,6 @@
 #![allow(missing_docs, reason = "integration test")]
 
-use pair::{HasDependent, Owner, Pair};
+use pair::{Dependent, HasDependent, Owner, Pair};
 use std::convert::Infallible;
 
 #[derive(Debug)]
@@ -14,10 +14,7 @@ impl Owner for SimpleOwner {
     type Context<'a> = ();
     type Error = Infallible;
 
-    fn make_dependent(
-        &self,
-        (): Self::Context<'_>,
-    ) -> Result<<Self as HasDependent<'_>>::Dependent, Self::Error> {
+    fn make_dependent(&self, (): Self::Context<'_>) -> Result<Dependent<'_, Self>, Self::Error> {
         Ok(&self.0)
     }
 }
@@ -43,10 +40,7 @@ impl Owner for PairOwner {
     type Context<'a> = ();
     type Error = Infallible;
 
-    fn make_dependent(
-        &self,
-        (): Self::Context<'_>,
-    ) -> Result<<Self as HasDependent<'_>>::Dependent, Self::Error> {
+    fn make_dependent(&self, (): Self::Context<'_>) -> Result<Dependent<'_, Self>, Self::Error> {
         Ok(PairOwnerDependent {
             value_ref: &self.value,
             inner_pair_ref: &self.inner_pair,