+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 Implemented the core functionality of this crate:
 - Defined `Pair` struct
 - Defined `Owner` and `HasDependent` traits
+- Defined `Dependent` type alias
 
 ## v0.1.0